+++ ack-service/src/main/java/org/ack/service/ProductService.java
@@ -3,6 +3,6 @@ package org.ack.service;
 import org.ack.base.service.AckMapperService;
 import org.ack.pojo.Product;
 
-public interface ProductService extends AckMapperService<Product, Long>{
+public interface ProductService extends AckMapperService<Product, Integer>{
 
 }
